Melhoramento de Respostas e Criaçao da Camada de Servico

+++ src/main/java/com/viana/socialbooks/repoitory/LivrosRepository.java
@@ -12,7 +12,7 @@ import com.viana.socialbooks.domain.Livro;
  *
  * @Criado em 24 de nov de 2018
  */
-public interface LivrosRepository extends JpaRepository<Livro, Long>{
+public interface LivrosRepository extends JpaRepository<Livro, Integer>{
